requirements.txt
ninjapear
pydantic>=2.7.0
pydantic-ai
python-dotenv
httpx
